add the all the features. Moduler code error removed

+++ app.js
@@ -1,21 +1,42 @@
 
 const fs =require('fs');
+const chalk=require('chalk');
+const addTodo= (info)=>{
 
-const addTodo= (data)=>{
 
-
-    fs.writeFile("todo-data/data.json",JSON.stringify(data),(error)=>{
-        if(error){
-            console.log(error);
-        }else{
-            console.log("sucessfully");
+    fs.readFile("todo-data/data.json",(err,data)=>{
+        
+        if(data){
+            let arrData=Object.values(JSON.parse(data));
+            
+            arrData.push(info);
+            
+            fs.writeFile("todo-data/data.json",JSON.stringify(arrData),(error)=>{
+                if(error){
+                    console.log(chalk.red.bold.underline("Unable to add somthing went wrong"));
+                }else{
+                    
+                    console.table(arrData);
+                    console.log(chalk.green.bold.underline("Added Sucessfully"));
+                }
+            })
+        }else if(err){
+            let arr=[];
+            arr.push(info);
+            fs.writeFile("todo-data/data.json",JSON.stringify(arr),(error)=>{
+                if(error){
+                    console.log(chalk.red.bold.underline("Unable to add somthing went wrong"));
+                }else{
+                    console.log(chalk.green.bold.underline("Added Sucessfully"));
+                }
+            })
         }
     })
 }
 const showTodo=()=>{
     fs.readFile("todo-data/data.json",(error,data)=>{
         if(error){
-            console.log("File is Empty");
+            console.log(chalk.red.bold.underline("Data Empty"));
         }else{
             console.table(JSON.parse(data));      
 
@@ -50,7 +71,7 @@ const deleteTask=(task)=>{
                     if(error){
                         console.log(error);
                     }else{
-                        console.log("sucessfully");
+                        console.info(chalk.green.bold.underline("Deleted Sucessfully"));
                     }
                 });
                 console.table(info);
@@ -59,15 +80,20 @@ const deleteTask=(task)=>{
         }
     });
 }
-const deleteTODO=()=>{
-   
+const deleteTODO=(yOrNO)=>{
+    
+   if(yOrNO){
     fs.unlink("todo-data/data.json",(error)=>{
         if(error){
-            console.log(error);
+            console.info(chalk.red.bold.underline("Empty File"));
         }else{
-            console.log("sucessfully");
+            console.info(chalk.green.bold.underline(" Sucessfully Deleted All"));
         }
     });
+   }else{
+       process.exit();
+   }
+    
     
 }
                 
@@ -83,14 +109,15 @@ const updateTodo=(id,info)=>{
                 data=JSON.parse(data);
               for(var i=0;i<data.length;i++){
                     if(i==id){
-                        if(info.Task.length){
+                        if(info.Importance_Level){
+                            data[i].Importance_Level=info.Importance_Level;
+                        }
+                        if(info.Task){
                             data[i].Task=info.Task;
                         }
-                         if(info.ImportanceLevel){
-                            data[i].ImportanceLevel=info.ImportanceLevel;
-                        }
-                         if(info.TimeFrom){
-                            data[i].TimeFrom=info.TimeFrom;
+                         
+                         if(info.Time_From){
+                            data[i].Time_From=info.Time_From;
                         }
             
                     }
